添加登入页, 初步授权和注销完成

+++ src/main/java/net/fdxxw/oauth/entity/UserApp.java
@@ -9,9 +9,9 @@ import lombok.Data;
 @Data
 public class UserApp {
 
-    private Long userAppId;
+    private Integer userAppId;
 
-    private Long userId;
+    private Integer userId;
 
     private String appId;
 }
